% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.blocks.R
\name{add.blocks}
\alias{add.blocks}
\title{Adds a block structure to an incidence matrix}
\usage{
add.blocks(I, blocks = 2, density = 0.5)
}
\arguments{
\item{I}{matrix: An incidence matrix}

\item{blocks}{integer: number of blocks to add (between 2 and 26)}

\item{density}{numeric: desired within-block density}
}
\value{
matrix: An incidence matrix, row and column names begin with a letter indicating their block membership
}
\description{
\code{add.blocks} shuffles an incidence matrix to have a block structure or planted partition while preserving the row and column sums
}
\details{
Stochastic block and planted partition models generate graphs in which the probability that two nodes are connected
depends on whether they are members of the same or different blocks/partitions. Functions such as \link[igraph]{sample_sbm}
can randomly sample from stochastic block models with given probabilities. In contrast \code{add.blocks} adds a block
structure to an existing incidence matrix while preserving the row and column sums. Each row and each column are randomly
assigned to one of \code{blocks} number of groups, then marginal-perserving checkerboard swaps are performed that increase
the within-block density, until \code{density} is achieved (if possible).
}
\examples{
I <- incidence.from.probability(R = 100, C = 100, P = .1)
blockedI <- add.blocks(I, blocks = 2, density = .7)
all(rowSums(I)==rowSums(blockedI))
all(colSums(I)==colSums(blockedI))
}
\references{
{Neal, Z. P., Domagalski, R., and Sagan, B. 2021. Comparing alternatives to the fixed degree sequence model for extracting the backbone of bipartite projections. \emph{Scientific Reports, 11}, 23929. \doi{10.1038/s41598-021-03238-3}}
}
