% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.from.congress.R
\name{incidence.from.congress}
\alias{incidence.from.congress}
\title{Generate bill sponsorship incidence matrices and bipartite graphs}
\usage{
incidence.from.congress(
  session = NULL,
  types = NULL,
  sponsors = FALSE,
  names = TRUE,
  format = "data"
)
}
\arguments{
\item{session}{numeric: the session of congress (currently 108-117)}

\item{types}{vector: types of bills to include. May be any combination of c(\"s\", \"sres\", \"sjres\", \"sconres\") OR any combination of c(\"hr\", \"hres\", \"hjres\", \"hconres\").}

\item{sponsors}{boolean: should sponsors be distinguished from co-sponsors using edge weights}

\item{names}{boolean: should legislators be identified by name (TRUE), or by bioguideId (FALSE)}

\item{format}{string: format of output, one of c(\"data\", \"igraph\")}
}
\value{
If \code{format = "data"}, a list containing an incidence matrix, a dataframe of legislator characteristics, and a dataframe of bill characteristics.

If \code{format = "igraph"}, a bipartite igraph object composed of legislator vertices and bill vertices, each with vertex attributes.
}
\description{
\code{incidence.from.congress()} uses data from \href{https://www.congress.gov/}{https://www.congress.gov/} to construct an incidence
matrix or bipartite graph recording legislators' bill (co-)sponsorships.
}
